requirements.txt

future
emoji
lxml
beautifulsoup4
requests
sqlalchemy==1.3.20
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
faker
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
pyrogram>=1.0.7
tgCrypto>=1.2.2
fontTools
gtts
tswift
pytz
youtube_dl
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
telegraph
